refactor(plugin-docgen): include default context vars in blueprint params
Adds default context vars to the blueprint params
that are passed into plugins during the doc generation
process.
This is done to add context about the host application
and the version of the plugin framework, blueprint framework
and the plugin protocol versions being used by the doc gen tool.
This will help with providing context for issues related
to the plugin doc gen tool and how it interacts with plugins.
This same approach is used in the deploy engine and
the blueprint language server to provide context to issues
that point to specific versions of different components
that are used for doc generation, editor tools, validation
and deployments.