feat(blueprint): add stub resource for loading placeholder blueprints
Adds stub resource in the core provider to allow loading of blueprints
that have no real resources.
This is used as a work around for the design of the blueprint loader
and state container to be able to destroy blueprint instances without
requiring the user to provide a source blueprint document.
This is because the destroy functionality does not use or need
any of the data from a loaded blueprint as it operates with
a provided change set and the current state of a blueprint instance.