test(deploy-engine): remove dependency on external plugin binaries for integrated tests
Remove the dependency on external plugin binaries for
the integrated test suites.
It adds a lot of time and complexity to test runner script
to have to either source or build the test plugins for
the CI os and architecture to test the end-to-end
integration between plugins and the deploy engine.
As the test suites do not require any plugins to be
registered, this removes the dependency, trusting
the plugin framework test suites to provide sufficient
integrated testing of plugins.