chore(blueprint): update blueprint spec version for json with comments
Updates blueprint spec version to a new early stage
revision that supports JSON with Commas and Comments
to allow for a better user experience defining blueprints
with a JSON-based format.
This also includes 2 new dependencies that will be used
to support JWCC and allow for parsing line information
for JSON-based files to enable the language server
capabilities for JSON-based blueprints.