feat(blueprint): add allowed values to resource definitions schema
Adds an allowed values field to a resource definition schema
to allow for enum types for string, integer and float values
in a resource spec.
This includes validation that does strict checks when scalar
values are provided and produces warnings for string
interpolation where the value is not known until the substitution
value is resolved during change staging and deployment.