fix(blueprint): ensure annotations are resolved as scalar types other than strings
As per the blueprint specification, annotation values can be strings,
integers, floats or booleans. The implementation before this commit
would resolve all exact values as strings, other types would only be
resolved if a ${} substitution is used for an annotation value.
The changes included with this commit will resolve literal values
defined for annotations in a blueprint to the more precise scalar type.