fix(blueprint): add fix to ensure errors are separated from diags
Adds fix to ensure that errors are separated
from warnings and info diagnostics in the result
of the `ValidateLinkAnnotations` function so that the
error is not ignored by the validation process.
The validation process separates error froms other kinds
of diagnostics so that it is easier to evaluate
if validation has failed overall when loading a blueprint.