feat(blueprint): add custom validation for individual resource definition schema elements
Adds support for custom validation functions that can be defined
for individual resource definition schema elements.
This allows custom value-based validation and conditional validation
based on other values in the resource as defined in the source
blueprint.
This validation is limited to scalar types
(integers, floats, strings and booleans) and will not be called
for strings that contain ${..} substitutions.