feat(plugin-framework): enhance value setter util struct
This commit changes the `pluginutils.SpecValueSetter`
to `pluginutils.ValueSetter` that is a more general purpose
struct util for setting values on a target struct
based on an input mapping node.
This can be used for resource specs and to apply updates
in link implementations.
The new iteration of the value setter also takes optional
configuration that will only set values based on whether
the field has changed as determined by a pre-computed
list of modified fields.