feat(plugin-framework): add save operation interface and runner
Adds a helper for building plugins that follows
a prepare and execute pattern for multiple save operations
used to create or update a resource.
The prepare phase is for extracting data from a resolved
resource spec and producing structs that can be passed
into the upstream provider client library, this phase
will usually make use of the `ValueSetter` helper to
extract values from the spec data mapping node and write
them to a taregt struct. The execute phase will carry
out the actions with the upstream provider.