fix(lib-rt-sdk-python): upgrade pyo3 and fix approach to integrating axum with python coroutines
Due to changes to the pyo3 api and the introduction of pyo3-async-runtimes,
the previous approach for registering python coroutines (asyncio) as handlers
for the axum http server (on tokio multi-threaded runtime) no longer works.
A new approach using a background worker that calls the python async handlers
to process requests from the axum http server that communicates over channels
was introduced to get around the python structures such as TaskLocals that
can no longer be shared across thread boundaries safely due to the API changes.