test(runtime-libs): update java test client to use http 1.1
As the http/2 client doesn't seem to work even when the
underlying core rust axum server supports http/2, it reports
that there is an empty header for the http 1.1 parser even though
the client was set to http/2.
This seems to only be an issue on windows machines that was first
caught in the CI environment and was then replicated on a windows
development machine.