feat(lib-rt-ws-registry): add a best effort approach to message delivery in cluster
Adds a best effort implementation to provide a level
of resilience to the websocket registry when running in
a cluster where messages are broadcasted if the target
connection id for the message is not on the same node.
This adds an ack worker which uses an ephemeral store
to keep track of message delivery status between nodes
in a cluster with configurable timeouts to wait for
acknowledgements and maximum number of retry attempts
before the message should be considered lost.
This implementats the behaviour describe for the `celerity/api`
Celerity resource definition for WebSockets where the caller
can specify the clients to be informed if the message was lost
which will usually be the client that sent the message to be
shared with other clients that are a part of the same multi-user
session. The Celerity Binary Message Format for WebSockets
is used for the message sent to one or more clients
connected to the same node when a message is considered lost.