feat(lib-rt-consumer-redis): add redis consumer implementation
Adds a comprehensive redis consumer implementation
using redis streams with support for hiding messages
with message locks to be processed by one consumer
at a time, retry logic for message processing handlers,
optional behaviour to move messages to another redis
stream that acts like a dead-letter queue and
a maintenance worker task to trim the stream at an
interval with a shared locking mechanism for a best
effort approach to avoid multiple consumer nodes
in a cluster from trying to trim the stream at the
same time.
A best effort approach for stream trimming is
acceptable as the worst case is not critical,
either a strim is longer than expected for
a short period of time or multiple consumers
try to trim the stream which will lead to one
succeeding and the rest making uneccessary calls
for an already trimmed stream.